latest updates, mostly formatting twiddles from swift-format

+++ SwiftUIVizComponents/CircleDataPoint.swift
@@ -55,7 +55,7 @@ struct CircleDataPoint_Previews: PreviewProvider {
             ForEach(ColorScheme.allCases, id: \.self) { colorScheme in
                 ForEach([20, 40], id: \.self) { size in
                     PreviewBackground {
-                        CircleDataPoint(size: size,
+                        CircleDataPoint(size: CGFloat(size),
                                         position: CGPoint(x: 30, y: 30))
                             .previewDisplayName("\(size), \(colorScheme)")
                     } // PreviewBackground
