+++ SwiftUIVizComponents/DiamondDataPoint.swift
@@ -68,7 +68,7 @@ struct DiamondDataPoint_Previews: PreviewProvider {
             ForEach(ColorScheme.allCases, id: \.self) { colorScheme in
                 ForEach([20, 40], id: \.self) { size in
                     PreviewBackground {
-                        DiamondDataPoint(size: size,
+                        DiamondDataPoint(size: CGFloat(size),
                                          position: CGPoint(x: 30, y: 30))
                             .previewDisplayName("\(size), \(colorScheme)")
                     } // PreviewBackground
